requirements.txt
streamlit
pandas
scikit-learn
joblib
python-dotenv
google-generativeai